asf-3.15.0

+++ sam/applications/sam4l_qtouch_demo/ui.c
@@ -3,7 +3,7 @@
  *
  * \brief User Interface.
  *
- * Copyright (c) 2012 - 2013 Atmel Corporation. All rights reserved.
+ * Copyright (c) 2012 - 2014 Atmel Corporation. All rights reserved.
  *
  * \asf_license_start
  *
@@ -177,7 +177,7 @@ void ui_lcd_refresh_alphanum(bool ui_lcd_refresh,
 {
 	char  string_info[8];
 	if (ui_lcd_refresh == true ) {
-		sprintf(string_info, "%d", event_qtouch_slider_position);
+		sprintf(string_info, "%d", (int)event_qtouch_slider_position);
 		if (event_qtouch_slider_position < 100) {
 			string_info[2] = ' ';
 			string_info[3] = ' ';
